Shift load_bin_file into utils
